Refactor: Introduce `Mixins` for `GitHubRESTApi` (without breaking API :D)
